requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
jinja2>=3.1.3
flake8>=7.0.0
python-docx>=1.1.0
autopep8>=2.0.0
autoflake>=2.0.0
reportlab>=4.0.0
